+++ tests/testthat/test_core_api.R
@@ -8,13 +8,10 @@ test_that(
     val2 <- h3_is_valid(h3_address = c('whereami', '8abe8d12acaffff')),
     val3 <- h3_is_valid(h3_address = c('whereami', '8abe8d12acaffff'),
                         simple = FALSE),
-    expect_is(val1, 'logical'),
     expect_equal(val1, TRUE),
-    expect_is(val2, 'logical'),
     expect_equal(val2, c(FALSE, TRUE)),
     expect_is(val3, 'data.frame'),
     expect_is(val3$h3_address, 'character'),
-    expect_is(val3$h3_valid, 'logical'),
     expect_equal(val3$h3_valid, c(FALSE, TRUE))
   )
 )
@@ -26,11 +23,9 @@ test_that(
     expect_error(h3_is_pentagon(h3_address = 'whereami')),
     val1 <- h3_is_pentagon(h3_address = '8abe8d12acaffff'),
     val2 <- h3_is_pentagon(h3_address = '8abe8d12acaffff', simple = FALSE),
-    expect_is(val1, 'logical'),
     expect_equal(val1, FALSE),
     expect_is(val2, 'data.frame'),
     expect_is(val2$h3_address, 'character'),
-    expect_is(val2$h3_pentagon, 'logical'),
     expect_equal(val2$h3_pentagon, FALSE)
     # note to self find the addy of a pentagon
   )
@@ -43,11 +38,9 @@ test_that(
     expect_error(h3_is_rc3(h3_address = 'whereami')),
     val1 <- h3_is_rc3(h3_address = '8abe8d12acaffff'),
     val2 <- h3_is_rc3(h3_address = '8abe8d12acaffff', simple = FALSE),
-    expect_is(val1, 'logical'),
     expect_equal(val1, FALSE),
     expect_is(val2, 'data.frame'),
     expect_is(val2$h3_address, 'character'),
-    expect_is(val2$h3_rc3, 'logical'),
     expect_equal(val2$h3_rc3, FALSE)
   )
 )
@@ -59,11 +52,9 @@ test_that(
     expect_error(h3_get_base_cell(h3_address = 'whereami')),
     val1 <- h3_get_base_cell(h3_address = '8abe8d12acaffff'),
     val2 <- h3_get_base_cell(h3_address = '8abe8d12acaffff', simple = FALSE),
-    expect_is(val1, 'integer'),
     expect_equal(val1, 95L),
     expect_is(val2, 'data.frame'),
     expect_is(val2$h3_address, 'character'),
-    expect_is(val2$h3_base_cell, 'integer'),
     expect_equal(val2$h3_base_cell, 95L)
   )
 )
@@ -75,11 +66,9 @@ test_that(
     expect_error(h3_get_res(h3_address = 'whereami')),
     val1 <- h3_get_res(h3_address = '8abe8d12acaffff'),
     val2 <- h3_get_res(h3_address = '8abe8d12acaffff', simple = FALSE),
-    expect_is(val1, 'integer'),
     expect_equal(val1, 10L),
     expect_is(val2, 'data.frame'),
     expect_is(val2$h3_address, 'character'),
-    expect_is(val2$h3_res, 'integer'),
     expect_equal(val2$h3_res, 10L)
   )
 )
@@ -91,7 +80,6 @@ test_that('geo_to_h3 returns an appropriately structured data frame with single 
             val1 <- geo_to_h3(lon = 153.023503, lat = -27.468920, res = 15),
             val2 <- geo_to_h3(lon = 153.023503, lat = -27.468920, res = 15,
                               simple = FALSE),
-            expect_is(val1, 'character'),
             expect_equal(val1, '8fbe8d12acad2f3'),
             expect_is(val2, 'data.frame'),
             expect_equal(ncol(val2), 4),
@@ -99,22 +87,19 @@ test_that('geo_to_h3 returns an appropriately structured data frame with single 
             expect_equal(names(val2), c('X', 'Y', 'h3_res', 'h3_address')),
             expect_is(val2$X, 'numeric'),
             expect_is(val2$Y, 'numeric'),
-            expect_is(val2$h3_address, 'character'),
             # xy checks
             expect_equal(val2$h3_address, '8fbe8d12acad2f3')
           ))
 
 test_that('geo_to_h3 returns an appropriately structured data frame with multiple input',
           c(
-            val1 <- geo_to_h3(lon = 153.023503, lat = -27.468920, res = seq(15)),
-            val2 <- geo_to_h3(lon = 153.023503, lat = -27.468920, res = seq(15),
+            val1 <- geo_to_h3(lon = 153.023503, lat = -27.468920, res = seq(0,15)),
+            val2 <- geo_to_h3(lon = 153.023503, lat = -27.468920, res = seq(0,15),
                               simple = FALSE),
-            expect_is(val1, 'character'),
-            expect_equal(length(val1), 15L),
-            expect_equal(val1[15], '8fbe8d12acad2f3'),
+            expect_equal(val1[16], '8fbe8d12acad2f3'),
             expect_is(val2, 'data.frame'),
-            expect_equal(ncol(val2), 4),
-            expect_equal(nrow(val2), 15),
+            expect_equal(ncol(val2), 4L),
+            expect_equal(nrow(val2), 16L),
             expect_equal(names(val2), c('X', 'Y', 'h3_res', 'h3_address')),
             expect_is(val2$X, 'numeric'),
             expect_is(val2$Y, 'numeric'),
@@ -135,8 +120,6 @@ test_that('h3 to geo returns an appropriate dataset',
             expect_equal(ncol(val2), 3),
             expect_equal(nrow(val2), 1),
             expect_equal(names(val2), c('h3_address', 'h3_x', 'h3_y')),
-            expect_is(val2$h3_x, 'numeric'),
-            expect_is(val2$h3_y, 'numeric'),
             # xy checks
             expect_equal(val2$h3_x, 153.0239032),
             expect_equal(val2$h3_y, -27.46852938)